Refactor: Move libs to separate libraries folder
